docs: remove embedded screenshot from README.md
As per the PR checklist, any assets used by README.md should be outside the `metadata` directory. I don't know why but it could be because screenshots are redundant OR this directory is not included in the extension users download, which could break things. Either way, it's simpler to remove the screenshot from the README for now.